requirements.txt
appinst==2.1.2
apptools==4.2.1
atom==0.3.9
backports.ssl-match-hostname==3.2a3
blaze==0.7.0
Canopy==1.5.2
canopydebugger==0.1.1.dev0
canopydebugger-addon==0.1.1.dev0
casuarius==1.1
chaco==4.5.0
cloud==2.4.6
codetools==4.2.0
configobj==5.0.6
coverage==3.7.1
DataShape==0.4.2
docutils==0.12
ecdsa==0.13
enable==4.4.1
enaml==0.9.8
enclosure==0.2
encore==0.6.0
enstaller==4.8.1
envisage==4.4.0
esky==0.9.8.dev0
etsproxy==0.1.2
Examples==7.3
Fabric==1.10.2
faulthandler==2.4
feedparser==5.1.3
flake8==2.2.5
freetype==2.5.3
futures==2.2.0
graphviz==0.4.8
grits-client==0.1
gspread==0.3.0
httplib2==0.9.2
idle==2.7.3
into==0.1.3
invoke==0.12.0
ipython==2.3.1
Jinja2==2.7.3
jsonpickle==0.4.0
kernmagic==0.2.0
keyring==4.0
kiwisolver==0.1.3
libjpeg==7.0
libpng==1.6.12
libxml2==2.9.2
MarkupSafe==0.23
matplotlib==1.4.2
mccabe==0.3.1
memory-profiler==0.31
MKL==10.3
mock==1.0.1
nose==1.3.4
numpy==1.9.2
oauth2client==2.0.0
pandas==0.16.2
paramiko==1.16.0
pep8==1.6.2
PIL==1.1.7
ply==3.4
psutil==2.2.0
PsychoPy==1.82.1
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
pyaudio==0.2.4
pycrypto==2.6.1
pydub==0.15.0
pyface==4.4.0
pyflakes==1.0.0
pyglet==1.1.4
Pygments==2.0.2
pyparsing==2.0.2
PySide==1.2.2
pytest==2.8.5
python-dateutil==2.2
PythonDoc==2.7.3
pytz==2014.9
PyYAML==3.11
pyzmq==14.5.0
Qt==4.8.6
readline==6.2.1
requests==2.5.1
rsa==3.3
scipy==0.15.1
shiboken==1.2.2
six==1.10.0
ssl-match-hostname==3.4.0.2
supplement==0.5.dev0
sympy==0.7.6
tornado==4.0.2
traits==4.5.0
traits-enaml==0.2.1
traitsui==4.4.0
Unipath==1.1
wheel==0.24.0
wxPython==2.8.10.1
zeromq==4.0.5
